react-spa: refactor, upgrade and add vitest
