+++ templates/react-spa/.vscode/extensions.json
@@ -1,9 +1,4 @@
 {
-  "recommendations": [
-    "csstools.postcss",
-    "dbaeumer.vscode-eslint",
-    "esbenp.prettier-vscode",
-    "bradlc.vscode-tailwindcss"
-  ],
+  "recommendations": ["dbaeumer.vscode-eslint", "esbenp.prettier-vscode", "bradlc.vscode-tailwindcss"],
   "unwantedRecommendations": []
 }
