+++ templates/react-spa/package.json
@@ -13,30 +13,33 @@
     "fix": "esmate task fix"
   },
   "dependencies": {
-    "@esmate/shadcn": "^3.5.1",
-    "@tanstack/react-devtools": "^0.8.4",
-    "@tanstack/react-query": "^5.90.12",
-    "@tanstack/react-query-devtools": "^5.91.1",
-    "@tanstack/react-router": "^1.141.1",
-    "@tanstack/react-router-devtools": "^1.141.1",
+    "@esmate/shadcn": "^3.6.0",
+    "@tanstack/react-devtools": "^0.8.6",
+    "@tanstack/react-query": "^5.90.20",
+    "@tanstack/react-query-devtools": "^5.91.2",
+    "@tanstack/react-router": "^1.157.15",
+    "@tanstack/react-router-devtools": "^1.157.15",
     "react": "^19.2.3",
     "react-dom": "^19.2.3"
   },
   "devDependencies": {
-    "@esmate/eslint": "^1.2.1",
+    "@esmate/eslint": "^1.2.3",
     "@esmate/prettier": "^1.0.6",
     "@tailwindcss/vite": "^4.1.18",
-    "@tanstack/router-plugin": "^1.141.1",
-    "@types/node": "^24.10.3",
-    "@types/react": "^19.2.7",
+    "@tanstack/router-plugin": "^1.157.15",
+    "@types/node": "^24.10.9",
+    "@types/react": "^19.2.9",
     "@types/react-dom": "^19.2.3",
-    "eslint": "^9.39.1",
-    "esmate": "^2.1.3",
-    "prettier": "^3.7.4",
+    "@vitejs/plugin-react": "^5.1.2",
+    "eslint": "^9.39.2",
+    "esmate": "^2.1.4",
+    "jsdom": "^27.4.0",
+    "prettier": "^3.8.1",
     "tailwindcss": "^4.1.18",
     "typescript": "^5.9.3",
-    "vite": "^7.2.7",
+    "vite": "^7.3.1",
     "vite-tsconfig-paths": "^5.1.4",
+    "vitest": "^4.0.18",
     "web-vitals": "^5.1.0"
   },
   "tasks": {
