+++ templates/react-spa/prettier.config.js
@@ -6,5 +6,5 @@ export default defineConfig({
     tailwindFunctions: ["cn"],
     tailwindStylesheet: "src/assets/styles/global.css",
   },
-  ignores: ["src/routeTree.gen.ts"],
+  ignores: ["src/lib/tanstack/router.ts"],
 });
