--- templates/react-spa/src/config/devtools/index.ts
@@ -1,1 +1,0 @@
-export { TanStackDevTools } from "./devtools";
