--- templates/react-spa/src/config/query/context.ts
@@ -1,7 +1,0 @@
-import { QueryClient } from "@tanstack/react-query";
-
-export function getTanStackQueryContext() {
-  const queryClient = new QueryClient();
-
-  return { queryClient };
-}
