--- templates/react-spa/src/config/query/index.ts
@@ -1,2 +1,0 @@
-export { getTanStackQueryContext } from "./context";
-export { TanStackQueryProvider } from "./provider";
