--- templates/react-spa/src/config/web-vitals/index.ts
@@ -1,1 +1,0 @@
-export * from "./report-web-vitals";
