+++ templates/react-spa/src/lib/tanstack/index.ts
@@ -1,0 +1,3 @@
+export { TanStackApp } from "./app";
+export { TanStackDevTools } from "./devtools";
+export { queryClient } from "./query";
