+++ templates/react-spa/src/lib/tanstack/router.ts
@@ -8,9 +8,9 @@
 // You should NOT make any changes in this file as it will be overwritten.
 // Additionally, you should also exclude this file from your linter and/or formatter to prevent it from being checked or modified.
 
-import { Route as rootRouteImport } from './routes/__root'
-import { Route as IndexRouteImport } from './routes/index'
-import { Route as DemoTanstackQueryRouteImport } from './routes/demo.tanstack-query'
+import { Route as rootRouteImport } from './../../routes/__root'
+import { Route as IndexRouteImport } from './../../routes/index'
+import { Route as DemoTanstackQueryRouteImport } from './../../routes/demo.tanstack-query'
 
 const IndexRoute = IndexRouteImport.update({
   id: '/',
