+++ templates/react-spa/src/lib/web-vitals.ts
@@ -1,3 +1,7 @@
+// If you want to start measuring performance in your app, pass a function
+// to log results (for example: reportWebVitals(console.log))
+// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
+
 export function reportWebVitals(onPerfEntry?: () => void) {
   if (onPerfEntry && typeof onPerfEntry === "function") {
     import("web-vitals").then(({ onCLS, onINP, onFCP, onLCP, onTTFB }) => {
