+++ templates/react-spa/src/main.tsx
@@ -1,48 +1,14 @@
-import { createRouter, RouterProvider } from "@tanstack/react-router";
 import { StrictMode } from "react";
 import ReactDOM from "react-dom/client";
 
-import { getTanStackQueryContext, TanStackQueryProvider } from "./config/query/index.ts";
 import "@/assets/styles/global.css";
+import { TanStackApp } from "@/lib/tanstack";
+import { reportWebVitals } from "@/lib/web-vitals";
 
-import { reportWebVitals } from "./config/web-vitals";
-import { routeTree } from "./routeTree.gen";
+ReactDOM.createRoot(document.getElementById("app") as HTMLElement).render(
+  <StrictMode>
+    <TanStackApp />
+  </StrictMode>,
+);
 
-const queryContext = getTanStackQueryContext();
-
-// Create a new router instance
-const router = createRouter({
-  routeTree,
-  scrollRestoration: true,
-  defaultPreload: "intent",
-  defaultPreloadStaleTime: 0,
-  defaultStructuralSharing: true,
-  context: { ...queryContext },
-});
-
-// Register the router instance for type safety
-declare module "@tanstack/react-router" {
-  interface Register {
-    router: typeof router;
-  }
-}
-
-// Render the app
-const rootElement = document.getElementById("app");
-
-if (rootElement && !rootElement.innerHTML) {
-  const root = ReactDOM.createRoot(rootElement);
-
-  root.render(
-    <StrictMode>
-      <TanStackQueryProvider {...queryContext}>
-        <RouterProvider router={router} />
-      </TanStackQueryProvider>
-    </StrictMode>,
-  );
-}
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
 reportWebVitals();
