+++ templates/react-spa/vite.config.ts
@@ -11,9 +11,12 @@ export default defineConfig({
     tanstackRouter({
       target: "react",
       autoCodeSplitting: true,
+      generatedRouteTree: "src/lib/tanstack/router.ts",
     }),
     viteReact(),
     tailwindcss(),
-    tsconfigPaths(),
+    tsconfigPaths({
+      projects: ["./tsconfig.json"],
+    }),
   ],
 });
